refactor: split git.rs god-file into modules
