--- src/commands/git.rs
@@ -1,1694 +1,0 @@
-use clap::{Args, Subcommand};
-use serde::{Serialize, Serializer};
-
-use homeboy::core::git::{
-    self, CherryPickOptions, GitOutput, GithubFindOutput, GithubIssueOutput, GithubPrFleetOutput,
-    GithubPrOutput, GithubPrReadinessOutput, IssueCloseOptions, IssueCloseReason,
-    IssueCommentOptions, IssueCreateOptions, IssueEditOptions, IssueFindOptions, IssueState,
-    PrCommentMode, PrCommentOptions, PrCreateOptions, PrEditOptions, PrFindOptions, PrFleetOptions,
-    PrLandOptions, PrLandOutput, PrLandRefreshHelper, PrMergeabilityReconcileOptions,
-    PrMergeabilityReconcileOutput, PrPolicyDecision, PrPolicyMergeOptions, PrPolicyOpenOptions,
-    PrRefreshOptions, PrRefreshOutput, PrRefreshStrategy, PrState, PushOptions, RebaseOptions,
-};
-use homeboy::core::BulkResult;
-
-use crate::commands::version;
-
-use super::CmdResult;
-
-const BODY_STDIN_LIMIT_BYTES: u64 = 256 * 1024;
-
-#[derive(Args)]
-pub struct GitArgs {
-    #[command(subcommand)]
-    command: GitCommand,
-}
-
-#[derive(Subcommand)]
-enum GitCommand {
-    /// Show git status for a component
-    Status {
-        /// JSON input spec for bulk operations.
-        /// Use "-" for stdin, "@file.json" for file, or inline JSON string.
-        #[arg(long)]
-        json: Option<String>,
-
-        /// Component ID (non-JSON mode). When omitted, the component is
-        /// auto-detected from CWD via the registry or a portable
-        /// `homeboy.json`.
-        component_id: Option<String>,
-
-        /// Workspace path to operate on directly. Useful for unregistered
-        /// checkouts (CI runners, ad-hoc clones, worktrees).
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Commit changes (by default stages all, use flags for granular control)
-    Commit {
-        /// Component ID (optional if provided in JSON body or auto-detected
-        /// from CWD).
-        component_id: Option<String>,
-
-        /// Commit message or JSON spec (auto-detected).
-        /// Plain text: treated as commit message.
-        /// JSON (starts with { or [): parsed as commit spec.
-        /// @file.json: reads JSON from file.
-        /// "-": reads JSON from stdin.
-        spec: Option<String>,
-
-        /// Explicit JSON spec (takes precedence over positional)
-        #[arg(long, value_name = "JSON")]
-        json: Option<String>,
-
-        /// Commit message (CLI mode)
-        #[arg(short, long)]
-        message: Option<String>,
-
-        /// Commit only staged changes (skip automatic git add)
-        #[arg(long)]
-        staged_only: bool,
-
-        /// Stage and commit only these specific files
-        #[arg(long, num_args = 1.., conflicts_with = "exclude")]
-        files: Option<Vec<String>>,
-
-        /// Stage all files except these (mutually exclusive with --files)
-        #[arg(long, num_args = 1.., conflicts_with = "files")]
-        exclude: Option<Vec<String>>,
-
-        /// Explicit include list (repeatable)
-        #[arg(long, num_args = 1.., conflicts_with = "exclude", conflicts_with = "files")]
-        include: Option<Vec<String>>,
-
-        /// Workspace path to operate on directly. Useful for unregistered
-        /// checkouts (CI runners, ad-hoc clones, worktrees).
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Push local commits to remote
-    Push {
-        /// JSON input spec for bulk operations.
-        /// Use "-" for stdin, "@file.json" for file, or inline JSON string.
-        #[arg(long)]
-        json: Option<String>,
-
-        /// Component ID (non-JSON mode). When omitted, the component is
-        /// auto-detected from CWD via the registry or a portable
-        /// `homeboy.json`.
-        component_id: Option<String>,
-
-        /// Push tags as well
-        #[arg(long)]
-        tags: bool,
-
-        /// Use `--force-with-lease` for safe force-pushes (e.g. after a
-        /// rebase). Refuses to overwrite the remote if it has commits the
-        /// local ref hasn't seen. Plain `--force` is intentionally not
-        /// exposed.
-        #[arg(long)]
-        force_with_lease: bool,
-
-        /// Push to this remote URL directly instead of the configured upstream.
-        /// Prefer this without credentials plus --token; embedding tokens in
-        /// the URL can expose them in process listings.
-        #[arg(long, value_name = "URL")]
-        remote_url: Option<String>,
-
-        /// GitHub token to inject into --remote-url for this invocation.
-        /// Requires a https://github.com/... remote URL.
-        #[arg(long, value_name = "TOKEN", requires = "remote_url")]
-        token: Option<String>,
-
-        /// Explicit push refspec, e.g. HEAD:refs/heads/my-branch.
-        #[arg(long, value_name = "REFSPEC")]
-        refspec: Option<String>,
-
-        /// Clear GitHub Actions checkout's auth extraheader so URL auth wins.
-        #[arg(long)]
-        strip_extraheader: bool,
-
-        /// Workspace path to operate on directly. Useful for unregistered
-        /// checkouts (CI runners, ad-hoc clones, worktrees).
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Rebase the current branch onto another ref.
-    ///
-    /// Default (no `--onto`) rebases onto the current branch's tracked
-    /// upstream (`@{upstream}`), same semantics as `git pull --rebase`.
-    /// Git's default rebase drops commits whose patch-id matches a commit
-    /// already in upstream — squash-merged PRs are NOT dropped (different
-    /// patch-id); that case will land in a follow-up.
-    ///
-    /// On conflict, the operation returns a failed result with git's
-    /// stderr. Resolve manually, then re-run with `--continue` or
-    /// `--abort`.
-    Rebase {
-        /// Component ID. When omitted, auto-detected from CWD.
-        component_id: Option<String>,
-
-        /// Target ref to rebase onto. Defaults to the current branch's
-        /// tracked upstream (`@{upstream}`).
-        #[arg(long, value_name = "REF")]
-        onto: Option<String>,
-
-        /// Continue an in-progress rebase after manual conflict resolution.
-        /// Mutually exclusive with `--abort`.
-        #[arg(long, conflicts_with = "abort")]
-        r#continue: bool,
-
-        /// Abort an in-progress rebase and return to the pre-rebase state.
-        #[arg(long)]
-        abort: bool,
-
-        /// Workspace path to operate on directly.
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Cherry-pick one or more commits onto the current branch.
-    ///
-    /// Accepts SHAs, branch names, and ranges (`<a>..<b>`) as positional
-    /// args. Use `--pr <n>` to pick all commits from a GitHub PR via `gh`.
-    /// Both can be combined.
-    ///
-    /// On conflict, returns a failed result. Resolve manually, then
-    /// re-run with `--continue` or `--abort`.
-    CherryPick {
-        /// Component ID. When omitted, auto-detected from CWD.
-        #[arg(long, short)]
-        component_id: Option<String>,
-
-        /// Commit refs to pick: SHAs, branches, ranges (`<a>..<b>`).
-        /// Multiple positional args allowed.
-        #[arg(value_name = "REF")]
-        refs: Vec<String>,
-
-        /// Cherry-pick all commits from a GitHub PR (repeatable).
-        /// Resolved via `gh pr view <n> --json commits`.
-        #[arg(long, value_name = "NUMBER")]
-        pr: Vec<u64>,
-
-        /// Continue an in-progress cherry-pick after manual conflict
-        /// resolution. Mutually exclusive with `--abort`.
-        #[arg(long, conflicts_with = "abort")]
-        r#continue: bool,
-
-        /// Abort an in-progress cherry-pick.
-        #[arg(long)]
-        abort: bool,
-
-        /// Workspace path to operate on directly.
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Pull remote changes
-    Pull {
-        /// JSON input spec for bulk operations.
-        /// Use "-" for stdin, "@file.json" for file, or inline JSON string.
-        #[arg(long)]
-        json: Option<String>,
-
-        /// Component ID (non-JSON mode). When omitted, the component is
-        /// auto-detected from CWD via the registry or a portable
-        /// `homeboy.json`.
-        component_id: Option<String>,
-
-        /// Workspace path to operate on directly. Useful for unregistered
-        /// checkouts (CI runners, ad-hoc clones, worktrees).
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Create a git tag
-    Tag {
-        /// Component ID. When omitted, the component is auto-detected from
-        /// CWD via the registry or a portable `homeboy.json`.
-        component_id: Option<String>,
-
-        /// Tag name (e.g., v0.1.2)
-        ///
-        /// Defaults to v<component version> if not provided.
-        tag_name: Option<String>,
-
-        /// Tag message (creates annotated tag)
-        #[arg(short, long)]
-        message: Option<String>,
-
-        /// Workspace path to operate on directly. Useful for unregistered
-        /// checkouts (CI runners, ad-hoc clones, worktrees).
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Manage GitHub issues for a component
-    Issue(IssueArgs),
-    /// Manage GitHub pull requests for a component
-    Pr(PrArgs),
-}
-
-// ---------------------------------------------------------------------------
-// `git issue` subcommand tree
-// ---------------------------------------------------------------------------
-
-#[derive(Args)]
-pub struct IssueArgs {
-    #[command(subcommand)]
-    command: IssueCommand,
-}
-
-#[derive(Subcommand)]
-enum IssueCommand {
-    /// Create a new issue
-    Create {
-        /// Component ID
-        component_id: String,
-
-        /// Issue title
-        #[arg(short, long)]
-        title: String,
-
-        /// Issue body (markdown). Prefer --body-file for long content.
-        #[arg(short, long, conflicts_with = "body_file")]
-        body: Option<String>,
-
-        /// Read body from a file ("-" for stdin)
-        #[arg(long, value_name = "PATH")]
-        body_file: Option<String>,
-
-        /// Issue label (repeatable)
-        #[arg(short, long)]
-        label: Vec<String>,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        /// (for unregistered checkouts — CI runners, ad-hoc clones)
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Comment on an existing issue
-    Comment {
-        /// Component ID
-        component_id: String,
-
-        /// Issue number
-        #[arg(short, long)]
-        number: u64,
-
-        /// Comment body (markdown). Prefer --body-file for long content.
-        #[arg(short, long, conflicts_with = "body_file")]
-        body: Option<String>,
-
-        /// Read body from a file ("-" for stdin)
-        #[arg(long, value_name = "PATH")]
-        body_file: Option<String>,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Find issues matching filters (dedup primitive)
-    Find {
-        /// Component ID
-        component_id: String,
-
-        /// Exact title match
-        #[arg(short, long)]
-        title: Option<String>,
-
-        /// Required label (repeatable — all labels must be present)
-        #[arg(short, long)]
-        label: Vec<String>,
-
-        /// State filter: open (default), closed, all
-        #[arg(short, long, default_value = "open")]
-        state: String,
-
-        /// Max results (default 30)
-        #[arg(long, default_value_t = 30)]
-        limit: usize,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Close an existing issue with a typed reason
-    Close {
-        /// Component ID
-        component_id: String,
-
-        /// Issue number
-        #[arg(short, long)]
-        number: u64,
-
-        /// Close reason: completed (default) or not-planned. Use
-        /// `not-planned` to suppress re-filing by `homeboy issues reconcile`
-        /// — the GitHub-native signal for "we have decided not to fix this."
-        #[arg(short, long, default_value = "completed")]
-        reason: String,
-
-        /// Optional closing comment (markdown). Posted before the state
-        /// transition. Prefer --comment-file for long content.
-        #[arg(short, long, conflicts_with = "comment_file")]
-        comment: Option<String>,
-
-        /// Read closing comment from a file ("-" for stdin)
-        #[arg(long, value_name = "PATH")]
-        comment_file: Option<String>,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Edit an existing issue's title, body, or labels
-    Edit {
-        /// Component ID
-        component_id: String,
-
-        /// Issue number
-        #[arg(short, long)]
-        number: u64,
-
-        /// New title (optional)
-        #[arg(short, long)]
-        title: Option<String>,
-
-        /// New body (markdown). Prefer --body-file for long content.
-        #[arg(short, long, conflicts_with = "body_file")]
-        body: Option<String>,
-
-        /// Read body from a file ("-" for stdin)
-        #[arg(long, value_name = "PATH")]
-        body_file: Option<String>,
-
-        /// Add labels (repeatable)
-        #[arg(long = "add-label", value_name = "LABEL")]
-        add_labels: Vec<String>,
-
-        /// Remove labels (repeatable)
-        #[arg(long = "remove-label", value_name = "LABEL")]
-        remove_labels: Vec<String>,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-}
-
-// ---------------------------------------------------------------------------
-// `git pr` subcommand tree
-// ---------------------------------------------------------------------------
-
-#[derive(Args)]
-pub struct PrArgs {
-    #[command(subcommand)]
-    command: PrCommand,
-}
-
-#[derive(Subcommand)]
-enum PrCommand {
-    /// Create a new pull request
-    Create {
-        /// Component ID
-        component_id: String,
-
-        /// Base branch (target of the PR)
-        #[arg(short, long)]
-        base: String,
-
-        /// Head branch (source of the PR)
-        #[arg(short = 'H', long)]
-        head: String,
-
-        /// PR title
-        #[arg(short, long)]
-        title: String,
-
-        /// PR body (markdown). Prefer --body-file for long content.
-        #[arg(short = 'B', long, conflicts_with = "body_file")]
-        body: Option<String>,
-
-        /// Read body from a file ("-" for stdin)
-        #[arg(long, value_name = "PATH")]
-        body_file: Option<String>,
-
-        /// Open as draft
-        #[arg(long)]
-        draft: bool,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Edit an existing PR's title or body
-    Edit {
-        /// Component ID
-        component_id: String,
-
-        /// PR number
-        #[arg(short, long)]
-        number: u64,
-
-        /// New title
-        #[arg(short, long)]
-        title: Option<String>,
-
-        /// New body (markdown)
-        #[arg(short = 'B', long, conflicts_with = "body_file")]
-        body: Option<String>,
-
-        /// Read body from a file ("-" for stdin)
-        #[arg(long, value_name = "PATH")]
-        body_file: Option<String>,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Find PRs matching filters
-    Find {
-        /// Component ID
-        component_id: String,
-
-        /// Base branch filter
-        #[arg(short, long)]
-        base: Option<String>,
-
-        /// Head branch filter
-        #[arg(short = 'H', long)]
-        head: Option<String>,
-
-        /// State filter: open (default), closed, merged, all
-        #[arg(short, long, default_value = "open")]
-        state: String,
-
-        /// Max results (default 30)
-        #[arg(long, default_value_t = 30)]
-        limit: usize,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Explain PR merge readiness without attempting a merge
-    Readiness {
-        /// Component ID
-        component_id: String,
-
-        /// PR number
-        #[arg(short, long)]
-        number: u64,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Post a comment on a PR. Three modes:
-    ///
-    /// 1. Plain: no marker flags — a fresh comment is appended.
-    /// 2. Sticky single-section (#1334): `--key <k>` finds-or-updates the one
-    ///    comment tagged `<!-- homeboy:key=<k> -->`. The whole `--body` becomes
-    ///    the comment body.
-    /// 3. Sectioned (#1348): `--comment-key <outer> --section-key <inner>`
-    ///    merges `--body` into section `<inner>` of the shared comment tagged
-    ///    `<!-- homeboy:comment-key=<outer> -->`. Other sections are preserved.
-    ///    `--header` sets the line printed after the outer marker on new
-    ///    comments; `--footer` / `--footer-file` sets a block printed after
-    ///    the last section (e.g. a tooling-versions <details> block). Both
-    ///    are preserved from existing comments on merge when omitted.
-    ///    `--section-order` pins section ordering (CSV of keys); default is
-    ///    alphabetical.
-    ///
-    /// Modes 2 and 3 are mutually exclusive. `--key` with `--comment-key` or
-    /// `--section-key` is an error.
-    Comment {
-        /// Component ID
-        component_id: String,
-
-        /// PR number
-        #[arg(short, long)]
-        number: u64,
-
-        /// Comment body (markdown). Prefer --body-file for long content.
-        #[arg(short = 'B', long, conflicts_with = "body_file")]
-        body: Option<String>,
-
-        /// Read body from a file ("-" for stdin)
-        #[arg(long, value_name = "PATH")]
-        body_file: Option<String>,
-
-        /// Sticky whole-body key (mode 2, PR #1334).
-        /// Mutually exclusive with --comment-key / --section-key.
-        #[arg(short, long, conflicts_with_all = ["comment_key", "section_key"])]
-        key: Option<String>,
-
-        /// Sectioned mode: outer shared-comment key (mode 3, #1348).
-        /// Must be combined with --section-key.
-        #[arg(long, requires = "section_key")]
-        comment_key: Option<String>,
-
-        /// Sectioned mode: inner per-section key (mode 3, #1348).
-        /// Must be combined with --comment-key.
-        #[arg(long, requires = "comment_key")]
-        section_key: Option<String>,
-
-        /// Sectioned mode: optional header line written after the outer
-        /// marker on freshly-created shared comments (e.g.
-        /// "## Homeboy Results — `<component>`"). Existing comment headers
-        /// are preserved on merge.
-        #[arg(long, requires = "comment_key")]
-        header: Option<String>,
-
-        /// Sectioned mode: optional footer block written after the last
-        /// section (e.g. a tooling-versions <details> block). Existing
-        /// footers are preserved on merge when this is omitted; passing
-        /// --footer or --footer-file overwrites the preserved footer.
-        /// Mutually exclusive with --footer-file.
-        #[arg(long, requires = "comment_key", conflicts_with = "footer_file")]
-        footer: Option<String>,
-
-        /// Sectioned mode: read footer content from a file ("-" for stdin).
-        /// Mutually exclusive with --footer.
-        #[arg(long, requires = "comment_key", value_name = "PATH")]
-        footer_file: Option<String>,
-
-        /// Sectioned mode: CSV of section keys in desired order. Sections
-        /// listed here come first in the given order; others are appended
-        /// alphabetically. Example: `--section-order lint,test,audit`.
-        #[arg(long, requires = "comment_key", value_delimiter = ',')]
-        section_order: Option<Vec<String>>,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Report and optionally land a fleet of pull requests.
-    Fleet {
-        /// Component ID
-        component_id: String,
-
-        /// PR numbers or URLs.
-        #[arg(value_name = "PR")]
-        refs: Vec<String>,
-
-        /// Update stale PR branches where GitHub can do so safely.
-        #[arg(long)]
-        update_branches: bool,
-
-        /// Merge green, clean PRs. Without this flag the command is read-only.
-        #[arg(long)]
-        apply: bool,
-
-        /// Merge method: merge, squash, or rebase.
-        #[arg(long, default_value = "squash")]
-        merge_method: String,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Compare GitHub mergeability with local git merge-tree evidence.
-    ReconcileMergeability {
-        /// Component ID
-        component_id: String,
-
-        /// PR number
-        #[arg(short, long)]
-        number: u64,
-
-        /// Workspace path to discover the component from a portable homeboy.json
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Evaluate PR open/merge policy.
-    Policy(PrPolicyArgs),
-    /// Refresh a PR branch from its current base and report conflicts/checks.
-    Refresh {
-        /// Component ID
-        component_id: String,
-
-        /// PR number or GitHub pull request URL.
-        pr: String,
-
-        /// Update strategy. `auto` uses branch/pull rebase git config, falling
-        /// back to rebase.
-        #[arg(long, default_value = "auto", value_parser = ["auto", "rebase", "merge", "ff-only"])]
-        strategy: String,
-
-        /// Push the refreshed PR branch when the worktree is clean and checks pass.
-        /// Uses --force-with-lease for rebase safety; plain force is not exposed.
-        #[arg(long)]
-        push: bool,
-
-        /// Lightweight check command to run after a clean refresh. Repeatable.
-        /// Defaults to `git diff --check` when omitted.
-        #[arg(long = "check", value_name = "COMMAND")]
-        checks: Vec<String>,
-
-        /// Workspace path to discover the component from a portable homeboy.json.
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Land a train of ready PRs sequentially, pausing on the first blocker.
-    Land {
-        /// Repository as owner/repo or host/owner/repo.
-        repo: String,
-
-        /// PR numbers or URLs. URLs must point at the selected repo.
-        #[arg(value_name = "PR")]
-        prs: Vec<String>,
-
-        /// Merge method: merge, squash, or rebase.
-        #[arg(long, default_value = "squash", value_parser = ["merge", "squash", "rebase"])]
-        merge_method: String,
-
-        /// Delete the PR branch after merge.
-        #[arg(long)]
-        delete_branch: bool,
-
-        /// Inspect and report what would land without merging or refreshing.
-        #[arg(long)]
-        dry_run: bool,
-
-        /// Safe helper program used to refresh a dirty dependent PR.
-        /// Not run through a shell. Combine with --refresh-helper-arg.
-        #[arg(long, value_name = "PROGRAM")]
-        refresh_helper: Option<String>,
-
-        /// Argument for --refresh-helper. Supports {repo}, {number}, {url}, {head_sha}.
-        #[arg(
-            long = "refresh-helper-arg",
-            value_name = "ARG",
-            requires = "refresh_helper"
-        )]
-        refresh_helper_args: Vec<String>,
-
-        /// Retry merge after this many base-branch-modified races.
-        #[arg(long, default_value_t = 1)]
-        max_base_retries: usize,
-    },
-}
-
-#[derive(Args)]
-pub struct PrPolicyArgs {
-    #[command(subcommand)]
-    command: PrPolicyCommand,
-}
-
-#[derive(Subcommand)]
-enum PrPolicyCommand {
-    /// Evaluate whether Homeboy may create or update a proposed PR.
-    Open {
-        /// Component ID
-        component_id: String,
-
-        /// Policy file path (YAML or JSON)
-        #[arg(long, value_name = "PATH")]
-        policy: String,
-
-        /// Change source, e.g. autofix, deps, generated, release-prep, agent.
-        #[arg(long)]
-        source: Option<String>,
-
-        /// Base branch.
-        #[arg(long)]
-        base: Option<String>,
-
-        /// Head branch.
-        #[arg(long)]
-        head: Option<String>,
-
-        /// Head repository owner/name.
-        #[arg(long = "head-repo")]
-        head_repository: Option<String>,
-
-        /// Base repository owner/name.
-        #[arg(long)]
-        repository: Option<String>,
-
-        /// Changed file path. Repeatable.
-        #[arg(long = "file", value_name = "PATH")]
-        files: Vec<String>,
-
-        /// Read changed file paths from a newline-delimited file.
-        #[arg(long, value_name = "PATH")]
-        files_file: Option<String>,
-
-        /// Read changed files from the current git working tree.
-        #[arg(long)]
-        files_from_git: bool,
-
-        /// Workspace path to discover the component from a portable homeboy.json.
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-    /// Evaluate whether an existing PR is safe to merge; optionally merge it.
-    Merge {
-        /// Component ID
-        component_id: String,
-
-        /// Policy file path (YAML or JSON)
-        #[arg(long, value_name = "PATH")]
-        policy: String,
-
-        /// PR number.
-        #[arg(short, long)]
-        number: u64,
-
-        /// Author login override. Defaults to GitHub PR metadata.
-        #[arg(long)]
-        author: Option<String>,
-
-        /// Base branch override. Defaults to GitHub PR metadata.
-        #[arg(long)]
-        base: Option<String>,
-
-        /// Head branch override. Defaults to GitHub PR metadata.
-        #[arg(long)]
-        head: Option<String>,
-
-        /// Head repository owner/name override. Defaults to GitHub PR metadata.
-        #[arg(long = "head-repo")]
-        head_repository: Option<String>,
-
-        /// Base repository owner/name override. Defaults to component remote.
-        #[arg(long)]
-        repository: Option<String>,
-
-        /// Merge the PR when policy allows it.
-        #[arg(long)]
-        merge: bool,
-
-        /// Merge method: merge, squash, or rebase.
-        #[arg(long, default_value = "squash")]
-        merge_method: String,
-
-        /// Workspace path to discover the component from a portable homeboy.json.
-        #[arg(long, value_name = "PATH")]
-        path: Option<String>,
-    },
-}
-
-pub enum GitCommandOutput {
-    Single(GitOutput),
-    Bulk(BulkResult<GitOutput>),
-    Issue(GithubIssueOutput),
-    Pr(GithubPrOutput),
-    PrRefresh(PrRefreshOutput),
-    PrReadiness(GithubPrReadinessOutput),
-    Find(GithubFindOutput),
-    ReconcileMergeability(PrMergeabilityReconcileOutput),
-    Policy(PrPolicyDecision),
-    Fleet(GithubPrFleetOutput),
-    Land(PrLandOutput),
-}
-
-impl Serialize for GitCommandOutput {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let (variant, payload) = match self {
-            GitCommandOutput::Single(output) => ("single", serde_json::to_value(output)),
-            GitCommandOutput::Bulk(output) => ("bulk", serde_json::to_value(output)),
-            GitCommandOutput::Issue(output) => ("issue", serde_json::to_value(output)),
-            GitCommandOutput::Pr(output) => ("pr", serde_json::to_value(output)),
-            GitCommandOutput::PrRefresh(output) => ("pr_refresh", serde_json::to_value(output)),
-            GitCommandOutput::PrReadiness(output) => ("pr_readiness", serde_json::to_value(output)),
-            GitCommandOutput::Find(output) => ("find", serde_json::to_value(output)),
-            GitCommandOutput::ReconcileMergeability(output) => {
-                ("reconcile_mergeability", serde_json::to_value(output))
-            }
-            GitCommandOutput::Policy(output) => ("policy", serde_json::to_value(output)),
-            GitCommandOutput::Fleet(output) => ("fleet", serde_json::to_value(output)),
-            GitCommandOutput::Land(output) => ("land", serde_json::to_value(output)),
-        };
-
-        let mut payload = payload.map_err(serde::ser::Error::custom)?;
-        let Some(object) = payload.as_object_mut() else {
-            return Err(serde::ser::Error::custom(
-                "git command output payload must serialize as an object",
-            ));
-        };
-        object.insert(
-            "variant".to_string(),
-            serde_json::Value::String(variant.into()),
-        );
-        payload.serialize(serializer)
-    }
-}
-
-pub fn run(args: GitArgs, _global: &crate::commands::GlobalArgs) -> CmdResult<GitCommandOutput> {
-    match args.command {
-        GitCommand::Status {
-            json,
-            component_id,
-            path,
-        } => {
-            if let Some(spec) = json {
-                let output = git::status_bulk(&spec)?;
-                let exit_code = if output.summary.failed > 0 { 1 } else { 0 };
-                return Ok((GitCommandOutput::Bulk(output), exit_code));
-            }
-
-            let output = git::status_at(component_id.as_deref(), path.as_deref())?;
-            let exit_code = output.exit_code;
-            Ok((GitCommandOutput::Single(output), exit_code))
-        }
-        GitCommand::Commit {
-            component_id,
-            spec,
-            json,
-            message,
-            staged_only,
-            files,
-            exclude,
-            include,
-            path,
-        } => {
-            // Explicit --json flag always uses JSON mode
-            if let Some(json_spec) = json {
-                let output = git::commit_from_json(component_id.as_deref(), &json_spec)?;
-                return match output {
-                    git::CommitJsonOutput::Single(o) => {
-                        let exit_code = o.exit_code;
-                        Ok((GitCommandOutput::Single(o), exit_code))
-                    }
-                    git::CommitJsonOutput::Bulk(b) => {
-                        let exit_code = if b.summary.failed > 0 { 1 } else { 0 };
-                        Ok((GitCommandOutput::Bulk(b), exit_code))
-                    }
-                };
-            }
-
-            // Auto-detect: check if positional spec looks like JSON or is a plain message
-            let (inferred_message, json_spec) = match &spec {
-                Some(s) => {
-                    let trimmed = s.trim();
-                    // JSON indicators: starts with { or [, uses @file, or - for stdin
-                    let is_json = trimmed.starts_with('{')
-                        || trimmed.starts_with('[')
-                        || trimmed.starts_with('@')
-                        || trimmed == "-";
-                    if is_json {
-                        (None, Some(s.clone()))
-                    } else {
-                        // Treat as plain commit message
-                        (Some(s.clone()), None)
-                    }
-                }
-                None => (None, None),
-            };
-
-            // JSON mode if auto-detected
-            if let Some(json_str) = json_spec {
-                let output = git::commit_from_json(component_id.as_deref(), &json_str)?;
-                return match output {
-                    git::CommitJsonOutput::Single(o) => {
-                        let exit_code = o.exit_code;
-                        Ok((GitCommandOutput::Single(o), exit_code))
-                    }
-                    git::CommitJsonOutput::Bulk(b) => {
-                        let exit_code = if b.summary.failed > 0 { 1 } else { 0 };
-                        Ok((GitCommandOutput::Bulk(b), exit_code))
-                    }
-                };
-            }
-
-            // CLI flag mode - use inferred message or explicit -m flag
-            let final_message = inferred_message.or(message);
-            let mut resolved_files = files;
-            if resolved_files.is_none() {
-                resolved_files = include;
-            }
-
-            let options = git::CommitOptions {
-                staged_only,
-                files: resolved_files,
-                exclude,
-                amend: false,
-            };
-            let output = git::commit_at(
-                component_id.as_deref(),
-                final_message.as_deref(),
-                options,
-                path.as_deref(),
-            )?;
-            let exit_code = output.exit_code;
-            Ok((GitCommandOutput::Single(output), exit_code))
-        }
-        GitCommand::Push {
-            json,
-            component_id,
-            tags,
-            force_with_lease,
-            remote_url,
-            token,
-            refspec,
-            strip_extraheader,
-            path,
-        } => {
-            if let Some(spec) = json {
-                let output = git::push_bulk(&spec)?;
-                let exit_code = if output.summary.failed > 0 { 1 } else { 0 };
-                return Ok((GitCommandOutput::Bulk(output), exit_code));
-            }
-
-            let output = git::push_at(
-                component_id.as_deref(),
-                PushOptions {
-                    tags,
-                    force_with_lease,
-                    remote_url,
-                    token,
-                    refspec,
-                    strip_extraheader,
-                },
-                path.as_deref(),
-            )?;
-            let exit_code = output.exit_code;
-            Ok((GitCommandOutput::Single(output), exit_code))
-        }
-        GitCommand::Pull {
-            json,
-            component_id,
-            path,
-        } => {
-            if let Some(spec) = json {
-                let output = git::pull_bulk(&spec)?;
-                let exit_code = if output.summary.failed > 0 { 1 } else { 0 };
-                return Ok((GitCommandOutput::Bulk(output), exit_code));
-            }
-
-            let output = git::pull_at(component_id.as_deref(), path.as_deref())?;
-            let exit_code = output.exit_code;
-            Ok((GitCommandOutput::Single(output), exit_code))
-        }
-        GitCommand::Tag {
-            component_id,
-            tag_name,
-            message,
-            path,
-        } => {
-            // Derive tag from version if not provided
-            let final_tag = match tag_name {
-                Some(name) => name,
-                None => {
-                    // Need component_id to look up version
-                    let id = component_id.as_ref().ok_or_else(|| {
-                        homeboy::core::Error::validation_invalid_argument(
-                            "componentId",
-                            "Missing componentId",
-                            None,
-                            Some(vec![
-                                "Provide a component ID: homeboy git tag <component-id>"
-                                    .to_string(),
-                                "Or specify a tag name: homeboy git tag <component-id> <tag-name>"
-                                    .to_string(),
-                            ]),
-                        )
-                    })?;
-                    let (out, _) = version::show_version_output(id)?;
-                    format!("v{}", out.version)
-                }
-            };
-
-            let output = git::tag_at(
-                component_id.as_deref(),
-                Some(&final_tag),
-                message.as_deref(),
-                path.as_deref(),
-            )?;
-            let exit_code = output.exit_code;
-            Ok((GitCommandOutput::Single(output), exit_code))
-        }
-        GitCommand::Rebase {
-            component_id,
-            onto,
-            r#continue,
-            abort,
-            path,
-        } => {
-            let output = git::rebase_at(
-                component_id.as_deref(),
-                RebaseOptions {
-                    onto,
-                    continue_: r#continue,
-                    abort,
-                },
-                path.as_deref(),
-            )?;
-            let exit_code = output.exit_code;
-            Ok((GitCommandOutput::Single(output), exit_code))
-        }
-        GitCommand::CherryPick {
-            component_id,
-            refs,
-            pr,
-            r#continue,
-            abort,
-            path,
-        } => {
-            let output = git::cherry_pick_at(
-                component_id.as_deref(),
-                CherryPickOptions {
-                    refs,
-                    prs: pr,
-                    continue_: r#continue,
-                    abort,
-                },
-                path.as_deref(),
-            )?;
-            let exit_code = output.exit_code;
-            Ok((GitCommandOutput::Single(output), exit_code))
-        }
-        GitCommand::Issue(args) => run_issue(args),
-        GitCommand::Pr(args) => run_pr(args),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// `git issue` dispatch
-// ---------------------------------------------------------------------------
-
-fn run_issue(args: IssueArgs) -> CmdResult<GitCommandOutput> {
-    match args.command {
-        IssueCommand::Create {
-            component_id,
-            title,
-            body,
-            body_file,
-            label,
-            path,
-        } => {
-            let body = resolve_body(body, body_file)?.unwrap_or_default();
-            let output = git::issue_create(
-                Some(&component_id),
-                IssueCreateOptions {
-                    title,
-                    body,
-                    labels: label,
-                    path,
-                },
-            )?;
-            Ok((GitCommandOutput::Issue(output), 0))
-        }
-        IssueCommand::Comment {
-            component_id,
-            number,
-            body,
-            body_file,
-            path,
-        } => {
-            let body = resolve_body(body, body_file)?.ok_or_else(|| {
-                homeboy::core::Error::validation_invalid_argument(
-                    "body",
-                    "Comment body is required (--body or --body-file)",
-                    None,
-                    None,
-                )
-            })?;
-            let output = git::issue_comment(
-                Some(&component_id),
-                IssueCommentOptions { number, body, path },
-            )?;
-            Ok((GitCommandOutput::Issue(output), 0))
-        }
-        IssueCommand::Find {
-            component_id,
-            title,
-            label,
-            state,
-            limit,
-            path,
-        } => {
-            let state = parse_issue_state(&state)?;
-            let output = git::issue_find(
-                Some(&component_id),
-                IssueFindOptions {
-                    title,
-                    labels: label,
-                    state,
-                    limit,
-                    path,
-                },
-            )?;
-            Ok((GitCommandOutput::Find(output), 0))
-        }
-        IssueCommand::Close {
-            component_id,
-            number,
-            reason,
-            comment,
-            comment_file,
-            path,
-        } => {
-            let reason = parse_issue_close_reason(&reason)?;
-            let comment = resolve_body(comment, comment_file)?;
-            let output = git::issue_close(
-                Some(&component_id),
-                IssueCloseOptions {
-                    number,
-                    reason,
-                    comment,
-                    path,
-                },
-            )?;
-            Ok((GitCommandOutput::Issue(output), 0))
-        }
-        IssueCommand::Edit {
-            component_id,
-            number,
-            title,
-            body,
-            body_file,
-            add_labels,
-            remove_labels,
-            path,
-        } => {
-            let body = resolve_body(body, body_file)?;
-            let output = git::issue_edit(
-                Some(&component_id),
-                IssueEditOptions {
-                    number,
-                    title,
-                    body,
-                    add_labels,
-                    remove_labels,
-                    path,
-                },
-            )?;
-            Ok((GitCommandOutput::Issue(output), 0))
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// `git pr` dispatch
-// ---------------------------------------------------------------------------
-
-fn run_pr(args: PrArgs) -> CmdResult<GitCommandOutput> {
-    match args.command {
-        PrCommand::Create {
-            component_id,
-            base,
-            head,
-            title,
-            body,
-            body_file,
-            draft,
-            path,
-        } => {
-            let body = resolve_body(body, body_file)?.unwrap_or_default();
-            let output = git::pr_create(
-                Some(&component_id),
-                PrCreateOptions {
-                    base,
-                    head,
-                    title,
-                    body,
-                    draft,
-                    path,
-                },
-            )?;
-            Ok((GitCommandOutput::Pr(output), 0))
-        }
-        PrCommand::Edit {
-            component_id,
-            number,
-            title,
-            body,
-            body_file,
-            path,
-        } => {
-            let body = resolve_body(body, body_file)?;
-            let output = git::pr_edit(
-                Some(&component_id),
-                PrEditOptions {
-                    number,
-                    title,
-                    body,
-                    path,
-                },
-            )?;
-            Ok((GitCommandOutput::Pr(output), 0))
-        }
-        PrCommand::Find {
-            component_id,
-            base,
-            head,
-            state,
-            limit,
-            path,
-        } => {
-            let state = parse_pr_state(&state)?;
-            let output = git::pr_find(
-                Some(&component_id),
-                PrFindOptions {
-                    base,
-                    head,
-                    state,
-                    limit,
-                    path,
-                },
-            )?;
-            Ok((GitCommandOutput::Find(output), 0))
-        }
-        PrCommand::Readiness {
-            component_id,
-            number,
-            path,
-        } => {
-            let output = git::pr_readiness(Some(&component_id), number, path)?;
-            let exit = if output.readiness.mergeable { 0 } else { 1 };
-            Ok((GitCommandOutput::PrReadiness(output), exit))
-        }
-        PrCommand::Comment {
-            component_id,
-            number,
-            body,
-            body_file,
-            key,
-            comment_key,
-            section_key,
-            header,
-            footer,
-            footer_file,
-            section_order,
-            path,
-        } => {
-            let body = resolve_body(body, body_file)?.ok_or_else(|| {
-                homeboy::core::Error::validation_invalid_argument(
-                    "body",
-                    "Comment body is required (--body or --body-file)",
-                    None,
-                    None,
-                )
-            })?;
-
-            // --footer / --footer-file share resolve_body; clap guarantees at
-            // most one is set. `None` → preserve existing footer on merge.
-            let footer = resolve_body(footer, footer_file)?;
-
-            let mode = match (key, comment_key, section_key) {
-                (Some(k), None, None) => PrCommentMode::StickyWholeBody { key: k },
-                (None, Some(ck), Some(sk)) => PrCommentMode::Sectioned {
-                    comment_key: ck,
-                    section_key: sk,
-                    header,
-                    footer,
-                    section_order,
-                },
-                (None, None, None) => {
-                    // Header / footer / section_order without the pair — clap
-                    // already caught this via `requires = "comment_key"`, but
-                    // double-check.
-                    PrCommentMode::Fresh
-                }
-                // Remaining cases are impossible due to clap `requires` /
-                // `conflicts_with_all`, but keep the match exhaustive.
-                _ => unreachable!(
-                    "clap argument parsing should have rejected incompatible --key / --comment-key / --section-key combos"
-                ),
-            };
-
-            let output = git::pr_comment(
-                Some(&component_id),
-                PrCommentOptions {
-                    number,
-                    body,
-                    mode,
-                    path,
-                },
-            )?;
-            Ok((GitCommandOutput::Pr(output), 0))
-        }
-        PrCommand::Fleet {
-            component_id,
-            refs,
-            update_branches,
-            apply,
-            merge_method,
-            path,
-        } => {
-            if refs.is_empty() {
-                return Err(homeboy::core::Error::validation_missing_argument(vec![
-                    "at least one PR number or URL".to_string(),
-                ]));
-            }
-            let output = git::pr_fleet(
-                Some(&component_id),
-                PrFleetOptions {
-                    refs,
-                    update_branches,
-                    apply,
-                    merge_method,
-                    path,
-                },
-            )?;
-            let exit = if output.success { 0 } else { 1 };
-            Ok((GitCommandOutput::Fleet(output), exit))
-        }
-        PrCommand::ReconcileMergeability {
-            component_id,
-            number,
-            path,
-        } => {
-            let output = git::pr_reconcile_mergeability(
-                Some(&component_id),
-                PrMergeabilityReconcileOptions { number, path },
-            )?;
-            Ok((GitCommandOutput::ReconcileMergeability(output), 0))
-        }
-        PrCommand::Policy(args) => run_pr_policy(args),
-        PrCommand::Refresh {
-            component_id,
-            pr,
-            strategy,
-            push,
-            checks,
-            path,
-        } => {
-            let strategy = parse_pr_refresh_strategy(&strategy)?;
-            let output = git::pr_refresh(
-                Some(&component_id),
-                PrRefreshOptions {
-                    pr,
-                    strategy,
-                    push,
-                    checks,
-                    path,
-                },
-            )?;
-            let exit = if output.success { 0 } else { 1 };
-            Ok((GitCommandOutput::PrRefresh(output), exit))
-        }
-        PrCommand::Land {
-            repo,
-            prs,
-            merge_method,
-            delete_branch,
-            dry_run,
-            refresh_helper,
-            refresh_helper_args,
-            max_base_retries,
-        } => {
-            let output = git::land_prs(PrLandOptions {
-                repo,
-                prs,
-                merge_method,
-                delete_branch,
-                dry_run,
-                refresh_helper: refresh_helper.map(|program| PrLandRefreshHelper {
-                    program,
-                    args: refresh_helper_args,
-                }),
-                max_base_retries,
-            })?;
-            let exit = if output.summary.blocked > 0 { 1 } else { 0 };
-            Ok((GitCommandOutput::Land(output), exit))
-        }
-    }
-}
-
-fn parse_pr_refresh_strategy(value: &str) -> homeboy::core::Result<PrRefreshStrategy> {
-    match value {
-        "auto" => Ok(PrRefreshStrategy::Auto),
-        "rebase" => Ok(PrRefreshStrategy::Rebase),
-        "merge" => Ok(PrRefreshStrategy::Merge),
-        "ff-only" => Ok(PrRefreshStrategy::FfOnly),
-        _ => Err(homeboy::core::Error::validation_invalid_argument(
-            "strategy",
-            "strategy must be one of auto, rebase, merge, or ff-only",
-            Some(value.to_string()),
-            None,
-        )),
-    }
-}
-
-fn run_pr_policy(args: PrPolicyArgs) -> CmdResult<GitCommandOutput> {
-    match args.command {
-        PrPolicyCommand::Open {
-            component_id,
-            policy,
-            source,
-            base,
-            head,
-            head_repository,
-            repository,
-            mut files,
-            files_file,
-            files_from_git,
-            path,
-        } => {
-            if let Some(files_file) = files_file {
-                files.extend(read_lines_file(&files_file)?);
-            }
-            let output = git::evaluate_open_policy(PrPolicyOpenOptions {
-                component_id,
-                path,
-                policy_path: policy,
-                source,
-                base,
-                head,
-                head_repository,
-                repository,
-                files,
-                files_from_git,
-            })?;
-            let exit = if output.allowed { 0 } else { 1 };
-            Ok((GitCommandOutput::Policy(output), exit))
-        }
-        PrPolicyCommand::Merge {
-            component_id,
-            policy,
-            number,
-            author,
-            base,
-            head,
-            head_repository,
-            repository,
-            merge,
-            merge_method,
-            path,
-        } => {
-            let output = git::evaluate_merge_policy(PrPolicyMergeOptions {
-                component_id,
-                path,
-                policy_path: policy,
-                number,
-                author,
-                base,
-                head,
-                head_repository,
-                repository,
-                merge,
-                merge_method: Some(merge_method),
-            })?;
-            let exit = if output.allowed { 0 } else { 1 };
-            Ok((GitCommandOutput::Policy(output), exit))
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Small input helpers
-// ---------------------------------------------------------------------------
-
-/// Resolve a body argument from either inline `--body` or a file path.
-/// Returns `Ok(None)` if neither is set. Supports `-` for stdin.
-fn resolve_body(
-    inline: Option<String>,
-    file: Option<String>,
-) -> homeboy::core::Result<Option<String>> {
-    if let Some(body) = inline {
-        return Ok(Some(body));
-    }
-    let Some(path) = file else {
-        return Ok(None);
-    };
-
-    if path == "-" {
-        use std::io::Read;
-        let mut bytes = Vec::new();
-        std::io::stdin()
-            .take(BODY_STDIN_LIMIT_BYTES + 1)
-            .read_to_end(&mut bytes)
-            .map_err(|e| {
-                homeboy::core::Error::internal_io(
-                    format!("Failed to read body from stdin: {}", e),
-                    Some("stdin".into()),
-                )
-            })?;
-        if bytes.len() > BODY_STDIN_LIMIT_BYTES as usize {
-            return Err(homeboy::core::Error::validation_invalid_argument(
-                "body_file",
-                "stdin body exceeds the retained byte limit",
-                Some(format!("limit_bytes={BODY_STDIN_LIMIT_BYTES}")),
-                None,
-            ));
-        }
-        let buf = String::from_utf8(bytes).map_err(|e| {
-            homeboy::core::Error::internal_io(
-                format!("Failed to decode body from stdin as UTF-8: {}", e),
-                Some("stdin".into()),
-            )
-        })?;
-        return Ok(Some(buf));
-    }
-
-    let content = std::fs::read_to_string(&path).map_err(|e| {
-        homeboy::core::Error::internal_io(
-            format!("Failed to read body file: {}", e),
-            Some(path.clone()),
-        )
-    })?;
-    Ok(Some(content))
-}
-
-fn read_lines_file(path: &str) -> homeboy::core::Result<Vec<String>> {
-    let content = std::fs::read_to_string(path).map_err(|e| {
-        homeboy::core::Error::internal_io(
-            format!("Failed to read lines file: {}", e),
-            Some(path.to_string()),
-        )
-    })?;
-    Ok(content
-        .lines()
-        .map(str::trim)
-        .filter(|line| !line.is_empty())
-        .map(ToString::to_string)
-        .collect())
-}
-
-fn parse_issue_state(s: &str) -> homeboy::core::Result<IssueState> {
-    match s {
-        "open" => Ok(IssueState::Open),
-        "closed" => Ok(IssueState::Closed),
-        "all" => Ok(IssueState::All),
-        other => Err(homeboy::core::Error::validation_invalid_argument(
-            "state",
-            format!("Unknown issue state '{}'", other),
-            None,
-            Some(vec!["Use one of: open, closed, all".into()]),
-        )),
-    }
-}
-
-fn parse_issue_close_reason(s: &str) -> homeboy::core::Result<IssueCloseReason> {
-    // Accept both kebab-case (CLI ergonomic) and snake_case (matches GitHub
-    // GraphQL state_reason values for symmetry with `--json stateReason`).
-    match s {
-        "completed" => Ok(IssueCloseReason::Completed),
-        "not-planned" | "not_planned" => Ok(IssueCloseReason::NotPlanned),
-        other => Err(homeboy::core::Error::validation_invalid_argument(
-            "reason",
-            format!("Unknown close reason '{}'", other),
-            None,
-            Some(vec!["Use one of: completed, not-planned".into()]),
-        )),
-    }
-}
-
-fn parse_pr_state(s: &str) -> homeboy::core::Result<PrState> {
-    match s {
-        "open" => Ok(PrState::Open),
-        "closed" => Ok(PrState::Closed),
-        "merged" => Ok(PrState::Merged),
-        "all" => Ok(PrState::All),
-        other => Err(homeboy::core::Error::validation_invalid_argument(
-            "state",
-            format!("Unknown PR state '{}'", other),
-            None,
-            Some(vec!["Use one of: open, closed, merged, all".into()]),
-        )),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use clap::Parser;
-
-    #[derive(Parser)]
-    struct TestCli {
-        #[command(subcommand)]
-        command: GitCommand,
-    }
-
-    #[test]
-    fn push_override_flags_parse() {
-        let cli = TestCli::try_parse_from([
-            "git",
-            "push",
-            "homeboy",
-            "--remote-url",
-            "https://github.com/Extra-Chill/homeboy",
-            "--token",
-            "secret-token",
-            "--refspec",
-            "HEAD:refs/heads/autofix",
-            "--strip-extraheader",
-        ])
-        .expect("push flags parse");
-
-        match cli.command {
-            GitCommand::Push {
-                component_id,
-                remote_url,
-                token,
-                refspec,
-                strip_extraheader,
-                ..
-            } => {
-                assert_eq!(component_id.as_deref(), Some("homeboy"));
-                assert_eq!(
-                    remote_url.as_deref(),
-                    Some("https://github.com/Extra-Chill/homeboy")
-                );
-                assert_eq!(token.as_deref(), Some("secret-token"));
-                assert_eq!(refspec.as_deref(), Some("HEAD:refs/heads/autofix"));
-                assert!(strip_extraheader);
-            }
-            _ => panic!("expected push command"),
-        }
-    }
-
-    #[test]
-    fn push_token_requires_remote_url_at_parse_time() {
-        let err =
-            match TestCli::try_parse_from(["git", "push", "homeboy", "--token", "secret-token"]) {
-                Ok(_) => panic!("--token should require --remote-url"),
-                Err(err) => err,
-            };
-
-        assert!(err.to_string().contains("--remote-url"));
-    }
-
-    #[test]
-    fn pr_readiness_flags_parse() {
-        let cli = TestCli::try_parse_from([
-            "git",
-            "pr",
-            "readiness",
-            "homeboy",
-            "--number",
-            "5805",
-            "--path",
-            "/tmp/homeboy",
-        ])
-        .expect("pr readiness flags parse");
-
-        match cli.command {
-            GitCommand::Pr(PrArgs {
-                command:
-                    PrCommand::Readiness {
-                        component_id,
-                        number,
-                        path,
-                    },
-            }) => {
-                assert_eq!(component_id, "homeboy");
-                assert_eq!(number, 5805);
-                assert_eq!(path.as_deref(), Some("/tmp/homeboy"));
-            }
-            _ => panic!("expected pr readiness command"),
-        }
-    }
-}
